Add cassette fly check to Player.OnSquish
The cassette fly state is intended to make the player invincible, but at the moment squish checks are still made against the player. This can cause deaths in modded situations when the return path crosses e.g. block entities.

This path adds a check at the start of OnSquish to return immediately if the player is in the CassetteFly state.