Make `TryGotoBestFit` predicates' `out` params correct
Some predicates use out parameters to extract information from the
method's IL.
Previously there was no guarantee that the values of the out parameters
will be the ones from the match.

This PR invokes all the predicates again after finding a match to set
the values of the out parameters to ones from the match.