Remove unused `PhotosensitiveMode` setting
The setting was a leftover from when #836 was being developed,
and was forgotten about when moving it to the vanilla menu.

Its existence causes issues when saving the photosensitive mode
preferences, as it takes precedence over vanilla. If the setting is
changed in the vanilla options menu, modded settings must also be
saved for the change to be actually remembered.

Removing the setting should prevent this from happeming.