Fix third party triggers not starting dialog
Vanilla naturally calls OnEnter and OnStay in the same frame when
entering the trigger, which would mean that we don't need the
OnEnter method. However, SJ's "in filtration" uses a Trigger Trigger
from CrystallineHelper to start a dialog cutscene, which only calls
OnEnter and not OnStay.

Removing OnEnter made the dialog not appear and cause a TAS desync!